qdoc: Fix broken links generated by \sincelist
The combination of single-exec mode and outputting to subdirectories
(offline mode) made \sincelist generate href's to new classes and
functions that are missing the target subdirectory.

This happened because we did not pass the relative node to the
function(s) that generate that list - proper link resolution
requires a check if the page containing the \sincelist command
comes from a different doc module, and the relative node is
used for that.

Fixes: QTBUG-87802
Change-Id: I30024d651a857b15375ef7e35dfaa21f0193ec46
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 15ebf3327191335c2c6eac4114ccb0553dc34f63)