qt5_create_translation: Fix handling of directory dependencies
lupdate supports whole directories as sources. Anyhow, in this
case we cannot just declare the directory as dependency; it will
force lupdate to re-run every time and emit a warning:

warning MSB8064: Custom build for item "..." succeeded, but specified dependency "<the directory specified as a source>" does not exist. This may cause incremental builds to work incorrectly.

The patch checks if each specified source is a directory, and if so:
* Generates globbing expressions for the list of extensions to be used.
* Search the specified directory for matching files.
* Add those files to the list of dependencies to give to CMake instead
  of the directory itself.

The list of source file extensions is derived from lupdate defaults,
but can be tweaked by a new '-extensions' argument.

Original patch done by Chris Djali.

Fixes: QTBUG-86192
Pick-to: 5.15
Change-Id: Icc22be4f3ee096361ac319208261cdc8191d9adf
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>