CMake: Fix usage of qt_create_translation in a CTest
Commit 1b006ab3d8954 declared an input file generated by CMake as a
BYPRODUCT of the consuming file to fix a policy warning CMP0058.
Anyhow, this is technically incorrect, since the file is actually not produced,
but consumed by the lupdate command. The incorrect fix also had the
side-effect that ninja clean did remove the file, which prevents the use
of the macro in CTest.

Just remove the BYPRODUCT declaration. The file is written by
CMake, so it actually doesn't need neither to be listed in DEPENDENCIES
nor in BYPRODUCT.

Task-number: QTBUG-71354
Change-Id: I7e5677f2bbd84b8a9fffdcb73465dfae6012ceb5
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit 3101eae233f1f5031c5c814d70dedbc7986e7db8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>