lupdate: Fix assertion in C++ parser
The parser reserves space for the the string yyWord: the length of the
input string, e.g. the file.  Later, it checks via Q_ASSERT whether the
string is still attached to the initial buffer.

In Qt6 we cannot rely anymore on QString staying attached to it's buffer
for all operations. Therefore, the assertion cannot hold in all cases.

Change the assertion to something weaker: ensure that the buffer
size (the string's capacity) stays the same during the parser run.

Pick-to: 6.0 6.0.0
Fixes: QTBUG-88647
Change-Id: I62ef2598b7023bc5bbc44aa477ac3bca60d6f355
Reviewed-by: Friedemann Kleint <Friedemann.Kleint@qt.io>