qdoc: Fix breadcrumb generation for normal page nodes
If a \page uses a single \ingroup command and a matching \group exists,
QDoc is supposed to add a link to that group in the navigation bar
(breadcrumbs) for that \page. This was broken as a result of a changed
inheritance structure of nodes.

Fixes: QTBUG-89835
Change-Id: I915d7ae41866fa89fcd0b133b4827f2e3d55bd54
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 228be3bd697a70a31c26c26d6d21d7bd0aebceda)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>