qdoc: Fix handling of overloads when adopting functions between parents
When adopting a function node from one parent to another, QDoc did not
touch the linked list of overloads for the adopted function. This
caused situations where traversing the node tree, trying to find
functions for an aggregate (parent), yielded also overloads that are
not supposed to be found as they are not members of that parent.

One visible place was the classes-with-obsolete-members generated list,
where QDoc erraneously found obsolete overloads for a class that had
adopted function(s) from the global namespace.

To fix, remove an adoptee from the overload list of its primary function
(maintained by the original parent), and re-add it to the overload list
maintained by the new parent.

Refactor functions that manage linked lists; using recursion is
unnecessary.

Fixes: QTBUG-89003
Change-Id: Ibd215cd4b1d97003ac79fe02a71452e4a7a154db
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 550c3031c2fbafe89daec4046dd45946ee8a788b)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>