qdoc: Don't clear the warning count in single-exec mode
As the warning count and limit determine the exit code, in single-exec
mode we need to keep track of the number of warnings for all
documentation sets that are part of the build.

This assumes that each documentation set retrives the warning limit
from a shared location in the config.

Fixes: QTBUG-90867
Change-Id: I7588a2a945094a02543a2cbd51ec5e3865666b47
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 173ecc34a2e8cf33325703b381003799aaaa42cb)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>