qdoc: Fix regression in handling of global functions
Commit 550c3031 improved the handling of \relates command, but
introduced a regression; When relating a primary function (a
first in the linked list of overloads) to another aggregate,
the primary's nextOverload pointer was cleared. Subsequent
\fn commands trying to document other overloads then failed.

The correct way is to move the head of the linked list forward.
However, if there are no overloads, we do not want to delete the
list entirely, as the function (typically, a global one) must
be searchable from the original (global) scope.

Fixes: QTBUG-91244
Change-Id: If80fa93ac5231d0676c0809f1acfa97733f9f9e8
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>
(cherry picked from commit 026edd4bcefd6f0cffa72c2385407c28042b4b36)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>