Add more guidance for missing clang feature checks
The existing warning mentions that CMAKE_PREFIX_PATH or LLVM_INSTALL_DIR
needs to be set, or that the relevant libraries could not be found.
If users fix these problems, the warning still persists. It is not
obvious to the user that they also need to reset the relevant FEATURE_*
cache entries to force these checks to be re-evaluated. Update the
warning message so that this is now explicitly stated.

Pick-to: 6.0 6.1
Change-Id: I99fbe52e5c5523dee31fd9104978e04d00354c17
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>