Fix regression in lconvert handling empty translations
Fixes a regression introduced by commit b96fe95da00aca6b, which enabled
a sanity check that now failed to account for empty string.

The serialized format of a QString is such that '-1' indicates a null
string, otherwise it's <length><QChar>+. Since QChar is 2 bytes,
length should therefore always be an even number ... except for -1.

While at it, also preserve the difference between an empty string and
a null string.

Fixes: QTBUG-91558
Pick-to: 6.1
Change-Id: Iffbc6ee2c94b8363d2c1d91440022a77dbef4772
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>