qdoc: Stop auto-linking to section titles
The purpose of the auto-linking feature is to link strings in the
text that resemble class or function names to API reference. However,
QDoc did not discriminate between the types of the resolved link
targets. Therefore, writing

  \section1 QString

on a random \page would potentially redirect all autolinks to QString
to that section.

To fix, turn Node::Genus into attributes so we can do bitwise operations
on the genus, and introduce Node::API that includes all API reference
node genera.

Node::API is not assigned to nodes directly, but is used in searches
to match against any API reference nodes.

Pick-to: 6.0 6.1
Task-number: QTBUG-91141
Change-Id: Ic0a22b7105e1278662f5ae7d4132980200638225
Reviewed-by: Paul Wicking <paul.wicking@qt.io>