QDoc: Let default be qmldefault
To allow the repurposing of the \default command for the
more generic default value use-case, make the current
specialization explicit as \qmldefault.

This patch also introduces a stub implementation of the
\default command. This ensures that existing documentation
sets that depend on this command will not issue new warnings.
The patch is as such a prerequisite for the introduction of
the new \default command.

[ChangeLog][qdoc] QDoc's \default command should specify
a default value. To document a default QML property, use
the new \qmlproperty command.

Task-number: QTBUG-81525
Change-Id: I3ed38e8a6f5a431247a2cea4ea661c87dd97d1b0
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>