qdoc: Re-implement \default command
Marking a QML property as a default property is now done with \qmldefault
command; repurpose the \default command to set a default value for a
property.

The default value is visible in the detailed property documentation as
an extra 'default: <value>' attribute. The value is taken as-is from
the argument passed to the command.

Fixes: QTBUG-81525
Change-Id: I7a4395e6e96046facfc3d75cc62a3bd01d04935b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>