Simplify the generation of the .json file used by lupdate clang based parser
The compile_command.json file is used by the tool only for the includes it provides.
One entry per source file is not needed.
Also, the file name appearing in this compile_command.json is hard coded
to a dummy name.
The file in the compile_command.json file is not used by the tool.
If the source file on which the tool is run is present in the
compile_command.json file, and the dir is the same the dir the tool
was launched in, then the getFileName() function won't return a fullPath.
For testing lupdate, we want the fullpath to appear in the
expectedoutput.txt.

Change-Id: Id0f0c425be9013155f857653a28a563b4e7a5f42
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>