qdoc: Make quick3, qt3 tags quick3d, qt3d
The old regex tried to detect a trailing '3D'. Anyhow, the 3 was already
consumed in the greedy [a-z0-9]* expression. So we need to use a
positive lookbehind expression to check that a trailing 'D' is actually
following a '3'.

Fixes: QTBUG-92590
Change-Id: I147e222b2f380a5d21155dcda96199720a384cc9
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit a59f39c7ea1a292093c4db00c5507e32868fa323)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>