Revert "QDoc: Make Generator::fileBase static"
This reverts commit 3e67b5e4df84421b77f90cbf242abb4a3c4c93a5.

The original implementation where fileBase() was a virtual function,
overridden by HtmlGenerator, was needed as that generator is the
only one that conditionally appends '-obsolete' to output file names.

By using a shared static helper across all generators we broke that
feature.

Modify the revert to resolve conflicts and override fileBase() also
for WebXMLGenerator; .webxml output should not refer to files with
-obsolete suffix.

Pick-to: 6.0 6.1
Fixes: QTBUG-92478
Change-Id: I7a7e3ef333724c96624569dd27b4812196931fb0
Reviewed-by: Paul Wicking <paul.wicking@qt.io>