Qt Designer: Fix alignment of horizontal QTableWidget headers
The horizontal QTableWidget headers are centered by default.

This is a problem since QTableWidgetItem::textAlignment() returns 0
for the default value and the header items are presented as list
widget items in the editor which have default alignment.

Populate the header list in the item editor correctly.
Add a default value to the property manager (in the form
of an attribute) and add a reset function.

Also pass a default for saving.

Fixes: QTBUG-85513
Change-Id: I78b95515f829cfb280f19a674575c8bd91fbb948
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Jarek Kobus <jaroslaw.kobus@qt.io>
(cherry picked from commit 3ecd3398218f4bddebeb034b4914699132cd5288)