qdoc: \generatelist: Ensure we link to the main namespace reference
When there are multiple documentation modules that document items under
the same namespace, we incorrectly linked to the wrong one (a sub-page
generated for a single module). This happened when using

\generatelist namespaces

in a module that was not the one where the namespace is documented
but has a dependency to it; as is the case with the main Qt reference
(qtdoc).

Add documentation comments to make the code slightly more understandable,
and add a test case for both the main reference page and the sub-page
generated for a namespace.

Fixes: QTBUG-90540
Change-Id: Ie883ea53896d9d70a17bc01e072648c62a9b8aa3
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 99f57491a12c2aa778ab9c8c9d50549d65701694)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>