qdoc: \generatelist: Ensure we link to the main namespace reference
When there are multiple documentation modules that document items under
the same namespace, we incorrectly linked to the wrong one (a sub-page
generated for a single module). This happened when using

\generatelist namespaces

in a module that was not the one where the namespace is documented
but has a dependency to it; as is the case with the main Qt reference
(qtdoc).

Add documentation comments to make the code slightly more understandable,
and add a test case for both the main reference page and the sub-page
generated for a namespace.

Pick-to: 6.2
Fixes: QTBUG-90540
Change-Id: Ie883ea53896d9d70a17bc01e072648c62a9b8aa3
Reviewed-by: Paul Wicking <paul.wicking@qt.io>