qdoc: Drop 'module' attribute from the index for page nodes
Page nodes, regardless of the node subtype, have no relation to any C++
module. Having this attribute in the index caused QDoc to search for
such a module, and create one if one was not found. This was a waste of
resources, and for pure documentation modules (such as 'QDoc' - the
QDoc Manual), caused broken links to a non-existing module page
(qdoc-module.html).

Fixes: QTBUG-94480
Change-Id: I2f608eba21ae977d3baaebf40885098ed1fe8d7e
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 3e88b552a1aac807504fa85bf752e6bd8542b573)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>