qdoc: Clean up index file generation
This is a continuation of the work started in commit 3e88b552.

Stop adding a 'physical module name' for all nodes; this name is
directly related to the \inmodule and \inqmlmodule commands, and
is unnecessary for most node types. Enforce the module name only
for classes, structs, unions, namespaces and header files.

Drop redundant code, and stop writing module/group members to the
index as unnecessary; they don't need to be read back in.

Simplify example file name generation; base the filename always
on the project name, example nodes typically never used the
\inmodule command anyway.

Pick-to: 6.2
Change-Id: I53a12b51f305cce28f5eb061d87ab11c61f54f6c
Reviewed-by: Paul Wicking <paul.wicking@qt.io>