qt6_add_translations: Add OUTPUT_TARGETS argument
Consider a static library that a user wants to ship.
And consider embedding .qm files into a resource within this static
library.

Then the user must ship the object libraries that belong to the resource
too. In order to do that, users can specify a variable name in
OUTPUT_TARGETS that is filled with the OUTPUT_TARGETS of the inner
qt6_add_resource call.

Specifying OUTPUT_TARGETS for a target that is not a static library is a
no-op.

Task-number: QTBUG-95180
Change-Id: I3984ca0082c118fa7516dde252330028f85bf1f1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>
(cherry picked from commit c7e7b67382146dd91727c773db4e745d1b57e580)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>