Doc: Remove mentions of the scoping of \target / \l pairs
The documentation for the `\target` command states that a target "can be
linked *from the same document* [...]".
This statement is followed by an example that implies that
the linking must be done "from within the same QDoc comment".

Other than those assertions being contradictory, they are both
erroneous, as a target is supposed to be able to be linked to from any
other file in the same project and even from files in other projects,
given the correct configuration.

To avoid misleading the reader, both mentions of a required scoping for
a target-link pair were removed.

Additionally, a minor typo was resolved in `qdoc-guide.qdoc`.

Change-Id: Id5e84d0baf86f2b32b9cfea33aea82790ed82ab8
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
(cherry picked from commit 1cbe28aae8d744b725ad8d53887d8fb9b94a0be8)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>