qt6_add_lrelease: Generate initial .ts files with target language
If the .ts file that is to be handled by qt6_add_lrelease does not exist
yet, the function creates an initial file, which can be used by lupdate
as base line.

However, the initial file did not have a target language set, and it
will never be set by subsequent lupdate calls.

Add a custom command that provides the initial .ts file by calling
"lupdate -ts <filename>", which guesses the target language from the
file name.

Pick-to: 6.2
Change-Id: Id878b1197673f286e1d16fb191e08821e7b81ae5
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>