lupdate: Default to absolute locations for message-less .ts files
If the .ts file that is to be handled by the CMake command
qt6_add_lrelease does not exist yet, the function creates an initial
file, which can be used by lupdate as base line.

This file is devoid of any messages and the subsequent .ts file will
have no locations. Fix that by changing the default of message-less .ts
files to absolute.

For new files, lupdate already defaulted to absolute locations.
Doing that for .ts files without messages seems just consistent.

Pick-to: 6.2
Change-Id: I051761ad916af3affdd70e57e77cc245fd399de1
Reviewed-by: Qt CI Bot <qt_ci_bot@qt-project.org>
Reviewed-by: Joerg Bornemann <joerg.bornemann@qt.io>