QDoc: Improve report when a file is not copied during test regeneration
QDoc currently provides a way to be tested trough its custom test
harness `tst_generatedOutput`.

`tst_generatedOutput` test QDoc by using it to build a series of QDoc
projects, comparing the current output to one that was previously
considered correct.

When `tst_generatedOutput` is provided with the `-regenerate` command,
each QDoc project that has a test is built and any artifact that is
designated as being relevant is copied to the `expected_output`
directory.

When `tst_generatedOutput` fails to copy one of the specified file, a
small test report will be generated for the fail, such as the
following:

```
FAIL! : tst_generatedOutput::teletypeCommand() 'QFile::copy(m_outputDir->filePath(file), fileInfo.filePath())' returned FALSE. ()
   Loc: [qt5/qttools/tests/auto/qdoc/generatedoutput/tst_generatedoutput.cpp(204)]
```

The report does not contain any indication of which of the designated
artifacts is missing, even if `tst_generatedOutput` has that
information at the time of the failure.

Identifying which artifact is the cause of the fail in a test with
multiple files may be time-consuming or, otherwise, distracting.

To ease the developer experience when such a fail occurs, additional
information containing the designated path of the artifact are now
included in the test report. For example:

```
FAIL!  : tst_generatedOutput::teletypeCommand() 'QFile::copy(m_outputDir->filePath(file), fileInfo.filePath())' returned FALSE. (Failed to copy 'tt_command_applies_a_code_styling_to_its_content.html')
   Loc: [qt5/qttools/tests/auto/qdoc/generatedoutput/tst_generatedoutput.cpp(204)]
```

Change-Id: I33ccd413182d9da67b082524a220316f77e66d53
Reviewed-by: Paul Wicking <paul.wicking@qt.io>