qdoc: Ensure compatibility with boolean index attribute "related"
When the latest QDoc handles index files in 5.15, it crashes
during destruction. The reason is that the string cannot be
converted to int, always returns 0 in toInt function, and no
other related child can be added if the relatedNodes list has
one related node.

By being compatible with the old bool type, it ensures that QDoc
can identify the index files per version.

Pick-to: 6.2
Fixes: QTBUG-96408
Change-Id: I91210a47a7b66f09cd9f436c93892fefcba8f97b
Reviewed-by: Paul Wicking <paul.wicking@qt.io>