QDoc: Add a warning for tokens that are too long to be parsed
`Tokenizer` uses a fixed-size buffer when parsing the sources for
tokens.

When a token that is too long is encountered, all characters that do not
fit into the buffer are discarded and the parsing continues.

When this happens, misleading warnings may be issued by QDoc as the
sources that it sees are only partially correct.

For example, a comment-block that does not fit into the buffer might be
seen as invalid even when it is not.

To ease the debugging of those problems, a warning is now issued when a
character is read when the buffer is already full.

To avoid issuing the same warning for each character, `Tokenizer` now
employs a boolean flag that is set when such a warning is issued and
is reset when a new token is requested.

Resetting the flag when a new token is requested ensures that we report
all warnings of this type that are encountered during the parsing phase.

Some documentation for the new warning was added to the `qodc-warnings`
page.

Pick-to: 6.2
Change-Id: I2ca4d86c201a8009d3f1b6760ed8c28f4401e114
Reviewed-by: Paul Wicking <paul.wicking@qt.io>