Doc: Replace instances of `will` in qdoc-warnings.qdoc
The documentation for QDoc's warnings used a mixed style with regards to
expressing that a warning is reported by QDoc.

In particular, `QDoc will issue a warning` and `QDoc issues a warning`.

To improve the consistency of the exposition, all instances of `will
issue` were changed to `issues`.

`issues` was preferred as a choice as it provides a more assertive and
active style.

Similar usages of `will` were changed for the same reason.
For example, `will compare` -> `compares`.

Pick-to: 6.2
Change-Id: Ibedcae2beccac7e1bd48c6457b7083c111f32740
Reviewed-by: Paul Wicking <paul.wicking@qt.io>