QDoc: Add an option to inspect the intermediate Atom representation
When QDoc parses a block-comment, it produces an intermediate
representation of `Atom` types that is later used as the processing base
to generate the required output.

To ease the process of inspecting this representation for debugging or
learning purposes, a new command-line-option, `atoms-dump`, was added.

When `atoms-dump` is passed as an argument, QDoc will print a
human-readable `Atom`s structure for each block-comment that is parsed
during its execution.

The information is printed in `Doc::Doc`, the only call location for
`DocParser::parse`, along with information about the starting and ending
location of the relevant block-comment.

A previously unused function, `Text::dump`, centralizes the logic of
generating a human-readable form and the action of outputting it.

`Text::dump` was slightly modified to emphasize the relation between
consecutive atoms enclosed in a block-like structure, such as those
between a `ParaLeft` and `ParaRight` atom.

An unused function, `Atom::dump`, was removed.
The function duplicated parts of the code of `Text::dump`, but for a
single atom.

To avoid spreading the surface of side-effecting functions and to avoid
splitting the responsibility of formatting atoms in a
human-readable-form, the function was not used as part of `Text::dump`
and was, hence, removed as superfluous instead.

Fixes: QTBUG-97775
Change-Id: If2933139575f97640e87f27c8ab1dbcbf7d14171
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>