Doc: Add a notice about the nomenclature of files generated by QDoc
For many topic commands, QDoc generates a file that represents a page in
the documentation.

QDoc generally generates a filename based on the arguments of the
command and further normalizes the name itself.

This means that the file that is generated might not be named as the
argument in the documentation file itself.

This might be surprising sometimes, especially for commands like
`\page`, where the name is completely inserted by the user and is not
tied to something else in the code.

To try to avoid those surprises and to make it easier for a user to
understand which file will contain the contents that they produced, some
documentation about the nomenclature that QDoc uses for files generated
by the commands was added to the topic commands section of QDoc's
documentation.

Task-number: QTBUG-97174
Pick-to: 6.2
Change-Id: Iaa6c0c846b88c0d244c0221229ca2b739ab7dbdd
Reviewed-by: Paul Wicking <paul.wicking@qt.io>