qdoc: Improve link text formatting
QDoc attempts to style links to C++ functions in a specific way,
by moving function arguments and the enclosing parentheses
out of the link. This was done in a way that had unintended
consequences for all link texts that included an opening
parenthesis, whether the link target was a function or not.

As a fix;

* Increase the specificity of the regular expression
  used for identifying a function-like link text.
* Store the Node we're linking to in generator, so
  we can match the regular expression only to function
  targets, not all links. This is also a performance
  improvement.
* DocParser: Stop generating AutoLink atoms inside an
  open (explicit) link. This was useless as it split
  the input into multiple nested link->autolink atoms,
  and the nested ones are discarded/treated as strings
  in the generator anyway.
* Refactor related generator functions and remove
  obsolete debugging code.

Pick-to: 6.2
Task-number: QTBUG-97949
Change-Id: I89429a73f13aa9995a79dfd24a8e473c620d6bc6
Done-with: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>