qdoc: Improve auto-generation of breadcrumbs
QDoc already had the capability to automatically generate correct next-page
and previous-page links for pages listed in a table-of-contents (TOC)
structure.

Since we already parse the TOC table, store also the navigational parent
(page that's one level up in the TOC hierarchy), and use the parent
chain to construct the navigation bar (breadcrumbs) in the HTML generator.

Some projects may want to include the TOC page itself as part of the
breadcrumbs while others might not, so make that option configurable.

Fixes: QTBUG-98278
Change-Id: Iac63daeaed5f0e3d7703761a2530a4d7c09a2cea
Reviewed-by: Paul Wicking <paul.wicking@qt.io>