QDoc: Deprecate \oldcode and \newcode
`\oldcode` and `\newcode` are a pair of commands that are used together
to show changes in API usage.

Due to the low amount of use for this feature and the arguable utility
of it, the command-pair is now deprecated and marked for removal in Qt
6.4.

A warning is added when an `\oldcode` command is encountered to warn
the user of the commands deprecation and the expected removal in the
near future.

The manual for QDoc was updated to notify of the deprecation and
soon-to-be removal.

`tst_generatedOutput` was updated to avoid referring to the command-pair
and the touched output files were regenerated.

Task-number: QTBUG-98499
Change-Id: I5af6a9ff8571d51418cad9bfe2567b8401f4dcae
Reviewed-by: Paul Wicking <paul.wicking@qt.io>
Reviewed-by: Topi Reiniö <topi.reinio@qt.io>