qdoc: Allow parameters for the \include command
Allow an arbitrary number (up to nine) of extra parameters to the
\include command, and replace the matching \1..\9 placeholders in
the quoted content with strings passed as parameters.

This enables reusing oft-repeated content but still customizing
each occurrence.

The original second argument for \include (the snippet id tag)
was constructed by taking everything up until a newline - the
extra parameters must however each be enclosed in braces to
support whitespace within a parameter. Keep the old syntax
functional as it is widely used although not explicitly
documented.

[ChangeLog][qdoc] QDoc now allows extra parameters for the
\include command to customize the quoted content.

Pick-to: 6.3
Fixes: QTBUG-48742
Change-Id: I03c7ae02660f090971e3ed769a62ed5ee18557f0
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>