qdoc: Remove code for automatic module resolution
QDoc had code to resolve the module a class/header/namespace lives in
without requiring an explicit \inmodule command. This was based on the
path of the source file (declaration), and the code included a list of
module names that was out of date. The whole mechanism was poorly
designed, resulting in documentation warnings if the source tree had an
unexpected structure.

Pick-to: 6.3
Fixes: QTBUG-99740
Change-Id: Ib036aa31a06cfc8a742bda45b703164a3cd2c2f9
Reviewed-by: Kai Koehne <kai.koehne@qt.io>