qdoc: Remove code for automatic module resolution
QDoc had code to resolve the module a class/header/namespace lives in
without requiring an explicit \inmodule command. This was based on the
path of the source file (declaration), and the code included a list of
module names that was out of date. The whole mechanism was poorly
designed, resulting in documentation warnings if the source tree had an
unexpected structure.

Fixes: QTBUG-99740
Change-Id: Ib036aa31a06cfc8a742bda45b703164a3cd2c2f9
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
(cherry picked from commit 1f3f5696a8345f9d790bc867bc6210897d790753)
Reviewed-by: Qt Cherry-pick Bot <cherrypick_bot@qt-project.org>