qdoc: Simplify static initialization in CppCodeMarker::addMarkUp
`CppCodeMarker::addMarkUp` uses two static collections to define what is
considered a keyword or a primitive type when lexing code.

The initialization was done manually, hidden behind a branch to ensure
that it was executed only once, by copying
the values from some statically initialized arrays.

Supposedly, this was necessary at the time the code was written for a
lack of a suitable constructor for a `QSet`.

As static values are already initialized only once and the manual
initialization mirrors what `QSet`'s `inizializer_list` constructor does,
the code was modified to remove the branching and initialize the
collections in place.

Change-Id: Ic2ca86054970bd257ca7cbaba12c17de2bc177d3
Reviewed-by: Paul Wicking <paul.wicking@qt.io>