qdoc: Remove quotefunction from the codebase
The codebase for QDoc implements a command, 'quotefunction', that eases
quoting a specifying function from a currently being-quoted file.

The command is unused in Qt's documentation, is undocumented in the QDoc
manual and only appears as a string (but not a call to the command
itself), inside of the files used by `tst_generatedOutput`.

While, arguably, there might be some use to higher-level commands that
simplify common operations that might be done trough the use of the
`quotefromfile` family of commands, `quotefunction` itself is,
currently, no more than dead code, and is thus removed, knowing that it
can be reintroduced with ease should the need arise, as a mean to
simplify the upcoming changes to the quoting system.

`DocParser::slashed` is now removed as it was used only during the
processing of a 'quotefunction' command.

The machinery which supported this feature in the `CodeMarker` family of
classes, `functionBeginRegExp` and `functionEndRegExp`, was removed as
it is now unused.

The `quotefunction` string in the input data to `tst_generatedOutput`
was removed and the output was regenerated to take the change into
account.

Change-Id: Ied2b04e0f3f1a74ff60300e83e2a514de872b7cb
Reviewed-by: Paul Wicking <paul.wicking@qt.io>