qdoc: DocBook: Fix incorrect static_cast in generateHeader()
There was an incorrect cast from an Aggregate to a NamespaceNode
without checking whether we in fact have a node representing a
namespace.

This resulted in a segfault under certain conditions, typically when
generating output for a ProxyNode.

Add a test for proxy page generation that (sporadically) resulted in
a segfault before this fix. The test case itself is misusing QDoc
commands to force a generation of a proxy page; this corner case
will be addressed in a later patch to output a warning.

Pick-to: 6.3 6.2
Fixes: QTBUG-100471
Change-Id: I4efff050a6be9a1305d97403b33e9e270e95a842
Reviewed-by: Luca Di Sera <luca.disera@qt.io>
Reviewed-by: Paul Wicking <paul.wicking@qt.io>