CMake: Fix qt_create_translation to not remove .ts files on "clean"
When running "make clean", .ts files were deleted.  This must not
happen, because .ts files are source files.

Remove the .ts file from the OUTPUT of the custom command that updates
the .ts file.  Add a timestamp file instead that is touched upon
updating the .ts file.

In qt_add_translation we must depend on this timestamp file.  If users
call qt_add_translation, we don't depend on a timestamp file but
directly on the .ts file like before.

Pick-to: 5.15 6.2 6.3
Fixes: QTBUG-96549
Change-Id: If93a6d1b47a33b29dc92fb987da2afcf8d29e895
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>