Fix FindWrapLibClang.cmake run for Clang-based lupdate
When FindWrapLibClang.cmake runs for the Clang-based lupdate parser,
it fails because the generator expression __qt_clang_genex in
target_link_libraries is not explicitly quoted. According to cmake docs
(https://cmake.org/cmake/help/latest/command/target_link_libraries.html#target-link-libraries): "If the ... contains any ; characters, e.g. after evaluation of a
${list} variable, be sure to use an explicitly quoted argument "$<...>" so
that this command receives it as a single <item>"

Change-Id: I3f730b767d9b89f5c4aa19272aab03eb12f0edb2
Reviewed-by: Kai Koehne <kai.koehne@qt.io>
Reviewed-by: Alexey Edelev <alexey.edelev@qt.io>
Reviewed-by: Alexandru Croitor <alexandru.croitor@qt.io>